+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - The IATI activity scope is calculated based on the benefitting countries and
   included in the IATI XML download as applicable
+- The error summary is now shown correctly when adding new matched effort in the
+  application
 
 ## Release 140 - 2023-11-28